fix: addressed circular dependency causing vite issues
